extract license types and helpers and reuse them
